pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanoslides"
version = "0.1.0"
description = "CLI and library foundation for AI-powered slide generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "rich>=13.0",
  "typer>=0.12",
]

[project.scripts]
nanoslides = "nanoslides.cli.main:run"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]
